refactor(graphql): trim barrel to reduce client bundle surface (CHAOS-1236)
