+++ src/lib/__tests__/capacityForecast.test.ts
@@ -3,13 +3,15 @@ import type { Session } from "next-auth";
 
 // Mock dependencies BEFORE importing the module under test
 vi.mock("@/lib/auth", () => ({ auth: vi.fn() }));
-vi.mock("@/lib/graphql", () => ({
+vi.mock("@/lib/graphql/client", () => ({
   graphqlClient: { isEnabled: () => true, query: vi.fn(), request: vi.fn() },
+}));
+vi.mock("@/lib/graphql/capacityFetchers", () => ({
   getCapacityForecastViaGraphQL: vi.fn(),
 }));
 
 import { auth } from "@/lib/auth";
-import { getCapacityForecastViaGraphQL } from "@/lib/graphql";
+import { getCapacityForecastViaGraphQL } from "@/lib/graphql/capacityFetchers";
 import { getCapacityForecast } from "../api";
 
 // Helper to mock auth session
