+++ src/lib/graphql/index.ts
@@ -1,9 +1,6 @@
 /**
  * GraphQL module exports.
+ *
+ * Keep this barrel empty to avoid pulling server-only/heavy modules into
+ * client bundles through transitive re-exports.
  */
-
-export * from "./types";
-export * from "./client";
-export * from "./queries";
-export * from "./investmentFetchers";
-export * from "./capacityFetchers";
